test: use integers of different sizes, even on 32-bit
The test optim/cse-size fials on 32-bit because it needs
two integers of different size but uses int & long.
These two types have indeed different sizes on 64-bit
(LP64) but not on 32-bit (ILP32).

Fix this by using short & int.

Reported-by: Ramsay Jones <ramsay@ramsayjones.plus.com>

+++ validation/optim/cse-size.c
@@ -1,7 +1,7 @@
 static void foo(void)
 {
-	unsigned p = 0;
-	long x;
+	unsigned short p = 0;
+	int x;
 
 	for (;;)
 		if (p)
